Реализовал фукцию открытия формы загрузки фотографии, реализовал валидацию этой формы
