+++ js/main.js
@@ -1,6 +1,6 @@
 import generatePhotos from './small-photos.js';
 import photosData from './mocks.js';
-import form from './form.js';
+import form from './form/form.js';
 
 generatePhotos(photosData);
 form();
